requirements.txt
python-dotenv==0.21.1
redis==4.4.0
web3==5.31.1
